update to cope with Rails StrongParameters feature
This feature was introduced to curb security issues that arose by
passing the attacker controlled `params` object around carelessly
and allowing the attacker to control meant-to-be-private variables.
The design intends to be very specific about which keys are fine
and is mean as explicit opt-in.

This implementation works around that and simply allows everything
that starts with `openid.`, which is still more restrictive than
before.